Dynamic program for TSP (OutOfMemory to solve)

+++ src/graph/base/GraphXYReader.java
@@ -12,7 +12,7 @@ public class GraphXYReader extends GraphReader {
     protected void processLine(final String[] line) {
 	final Collection<Node> nodes = this.graph.getNodes();
 	final long newNodeId = nodes.size() + 1;
-	final NodeXY newNode = new NodeXY(newNodeId, Integer.parseInt(line[0]), Integer.parseInt(line[1]));
+	final NodeXY newNode = new NodeXY(newNodeId, Double.parseDouble(line[0]), Double.parseDouble(line[1]));
 	this.graph.addNode(newNode);
 	NodeXY otherNode;
 	// Perform the next procedure only after reading the file?
